Add ordered list functionality. Divid methods into separate files for clarity.
